// mobile/src/api/generated/types.ts
export interface components {
    schemas: {
        /** HealthOut */
        HealthOut: {
            /** Status */
            status: string;
            /** Debug */
            debug: boolean;
            /** Mocks */
            mocks: {
                [key: string]: unknown;
            };
        };
        /** TokenPairOut */
        TokenPairOut: {
            /** Access */
            access: string;
            /** Refresh */
            refresh: string;
            user: components["schemas"]["UserOut"];
        };
        /** UserOut */
        UserOut: {
            /** Id */
            id: number;
            /** Email */
            email: string | null;
            /** Phone */
            phone: string | null;
            /** Role */
            role: string;
            /** First Name */
            first_name: string;
            /** Last Name */
            last_name: string;
            /** Node Id */
            node_id?: string | null;
        };
        /** ErrorOut */
        ErrorOut: {
            /** Detail */
            detail: string;
        };
        /** RegisterIn */
        RegisterIn: {
            /**
             * Email
             * Format: email
             */
            email: string;
            /** Password */
            password: string;
            /**
             * Role
             * @default node_manager
             */
            role: string;
            /**
             * First Name
             * @default
             */
            first_name: string;
            /**
             * Last Name
             * @default
             */
            last_name: string;
            /** Node Id */
            node_id?: string | null;
        };
        /** LoginIn */
        LoginIn: {
            /**
             * Email
             * Format: email
             */
            email: string;
            /** Password */
            password: string;
        };
        /** AccessTokenOut */
        AccessTokenOut: {
            /** Access */
            access: string;
        };
        /** RefreshIn */
        RefreshIn: {
            /** Refresh */
            refresh: string;
        };
        /** OTPRequestOut */
        OTPRequestOut: {
            /** Sent */
            sent: boolean;
            /** Mock */
            mock: boolean;
            /** Dev Code */
            dev_code?: string | null;
        };
// ...
    };
    responses: never;
    parameters: never;
    requestBodies: never;
    headers: never;
    pathItems: never;
}